+++ options.go
@@ -67,9 +67,12 @@ func WithRequestsType(requestsType string) Option {
 	}
 }
 
-// WithConnectionString accept connection string like 'grpc[s]://{endpoint}/?database={database}'
+// WithConnectionString accept connection string like
+//
+//   grpc[s]://{endpoint}/?database={database}
+//
 // Warning: WithConnectionString will be removed at next major release
-// (connection string will be required string param of ydb.New)
+// (connection string will be required string param of ydb.Open)
 func WithConnectionString(connectionString string) Option {
 	return func(ctx context.Context, c *connection) error {
 		var (