+++ test/connection_test.go
@@ -62,9 +62,9 @@ func TestConnection(t *testing.T) {
 	)
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*10)
 	defer cancel()
-	db, err := ydb.New(
+	db, err := ydb.Open(
 		ctx,
-		ydb.WithConnectionString(os.Getenv("YDB_CONNECTION_STRING")),
+		os.Getenv("YDB_CONNECTION_STRING"),
 		ydb.WithAnonymousCredentials(),
 		ydb.With(
 			config.WithOperationTimeout(time.Second*2),